moving fetchinator into own file
